ai-service/requirements.txt
flask
flask-cors
prophet
pandas
numpy
mysql-connector-python
scikit-learn
python-dotenv